add listassetindexassets RPC
Former-commit-id: 6d34c610a324715025bb437c0bfd1d505b75b7c9

+++ src/services/assetallocation.cpp
@@ -165,7 +165,7 @@ void WriteAssetIndexForAllocationAddress(const CAssetAllocationTuple& allocation
         std::vector<uint256> TXIDS;
         if(passetindexdb->ReadIndexTXIDs(allocationTuple, page, TXIDS)){
             // new page needed
-            if(TXIDS.size() >= fAssetIndexPageSize){
+            if(((int)TXIDS.size()) >= fAssetIndexPageSize){
                 TXIDS.clear();
                 page++;
             }
@@ -185,7 +185,7 @@ void WriteAssetIndexForAllocationAddress(const CAssetAllocationTuple& allocation
         std::vector<uint256> TXIDS;
         if(passetindexdb->ReadIndexTXIDs(allocationTuple.nAsset, page, TXIDS)){
             // new page needed
-            if(TXIDS.size() >= fAssetIndexPageSize){
+            if(((int)TXIDS.size()) >= fAssetIndexPageSize){
                 TXIDS.clear();
                 page++;
             }
